+++ lib/src/_core/utils/omni_video_player_controls_overlay.dart
@@ -313,7 +313,7 @@ class _OmniVideoPlayerControlsOverlayState
         onStartInteraction: onStartInteraction,
         onEndInteraction: onEndInteraction,
       ),
-      if (ctrl.isSeeking)
+      if (ctrl.isSeeking && !ctrl.isFinished)
         Positioned.fill(
           child: Align(
             alignment: Alignment.center,