+++ lib/src/_others/generic_playback_controller.dart
@@ -367,13 +367,12 @@ class GenericPlaybackController extends OmniPlaybackController {
       return;
     }
 
-    isSeeking = true;
-    _videoStuckCounter = 0;
-    _lastVideoPosition = Duration.zero;
-
     if (!_isSeeking) {
       wasPlayingBeforeSeek = isPlaying;
     }
+    isSeeking = true;
+    _videoStuckCounter = 0;
+    _lastVideoPosition = Duration.zero;
 
     if (position.inMicroseconds != 0 && !skipHasPlaybackStarted) {
       _hasStarted = true;
@@ -537,11 +536,9 @@ class GenericPlaybackController extends OmniPlaybackController {
 
   @override
   Future<void> replay({bool useGlobalController = true}) async {
-    await Future.wait([
-      pause(useGlobalController: useGlobalController),
-      seekTo(Duration.zero),
-      play(useGlobalController: useGlobalController),
-    ]);
+    await pause(useGlobalController: useGlobalController);
+    await seekTo(Duration.zero);
+    await play(useGlobalController: useGlobalController);
   }
 
   @override