+++ lib/src/_vimeo/vimeo_initializer.dart
@@ -53,7 +53,6 @@ class VimeoInitializer implements IOmniVideoPlayerInitializerStrategy {
             controller: controller,
             preferredQualities:
                 options.videoSourceConfiguration.preferredQualities,
-            autoPlay: options.videoSourceConfiguration.autoPlay,
           ),
           Container(color: Colors.transparent),
         ],