+++ lib/src/_youtube/youtube_webview_controller.dart
@@ -399,11 +399,9 @@ class YouTubeWebViewController extends OmniPlaybackController {
 
   @override
   Future<void> replay({bool useGlobalController = true}) async {
-    await Future.wait([
-      pause(useGlobalController: useGlobalController),
-      seekTo(Duration.zero),
-      play(useGlobalController: useGlobalController),
-    ]);
+    await pause(useGlobalController: useGlobalController);
+    await seekTo(Duration.zero);
+    await play(useGlobalController: useGlobalController);
   }
 
   @override